requirements.txt
Django==4.2.3
django-ckeditor==6.7.0
sqlparse==0.4.4
psycopg2-binary==2.9.6
Pillow==10.0.0